fixes for record

+++ src/main/java/PctCompressor.java
@@ -1,11 +1,11 @@
 public class PctCompressor {
 	private static final short PRECISION = 1000;
 
-	public static short compress(float pct) {
+	public static short compress(double pct) {
 		return (short) (pct * PRECISION);
 	}
 
-	public static float uncompress(short pctCompressed) {
-		return (float) (pctCompressed / PRECISION);
+	public static double uncompress(short pctCompressed) {
+		return  (double) pctCompressed / PRECISION;
 	}
 }
